// src/api/updateChannelSettingsRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Tubester API
 * To access protected endpoints, first log in:

[🔐 read only Login with Google](/api/auth/login/google?returnUrl=/swagger/index.html)

[🔐 write Login with Google](/api/auth/login/google/write?returnUrl=/swagger/index.html)
 * OpenAPI spec version: v1
 */

export interface UpdateChannelSettingsRequest {
  isCommentAssistantEnabled?: boolean;
  isSuggestRepliesForTopLevelCommentsOnly?: boolean;
  /**
   * @minimum 0
   * @maximum 2147483647
   */
  maxSuggestedRepliesPerSync?: number;
  /**
   * @minimum 0
   * @maximum 2147483647
   */
  maxCommentAgeDays?: number;
  /**
   * @minLength 1
   * @maxLength 50
   */
  replyLanguage: string;
  /**
   * @maxLength 500
   * @nullable
   */
  responseForNonTextualComments?: string | null;
}